refactor(shared): move DestinationDetailsSchema to shared, break send↔transactions/agicash-db cycles
DestinationDetailsSchema is a pure (zod-only) value-object for a send's destination (Agicash contact vs LN address). It lived in send/cashu-send-quote.ts, but the foundational agicash-db json-model and the transactions detail mapper both import it as a runtime value, making the DB/transactions layers depend up into send (the send↔agicash-db and send↔transactions cycles). Relocate it to shared/send-destination.ts; consumers import the schema from shared, and cashu-send-quote.ts re-exports the DestinationDetails type so send-internal call sites are unchanged. Pure relocation, no behavior change.

Part of the wallet-SDK no-cache migration (cycle-break before move).
