+++ apps/web-wallet/app/features/agicash-db/json-models/cashu-lightning-send-db-data.ts
@@ -1,6 +1,6 @@
 import { Money } from '@agicash/money';
 import { z } from 'zod/mini';
-import { DestinationDetailsSchema } from '~/features/send/cashu-send-quote';
+import { DestinationDetailsSchema } from '~/features/shared/send-destination';
 
 /**
  * Schema for cashu lightning send db data.
