+++ apps/web-wallet/app/features/transactions/transaction-details/cashu-lightning-send-transaction-details.ts
@@ -1,7 +1,7 @@
 import { Money } from '@agicash/money';
 import { z } from 'zod/mini';
 import { CashuLightningSendDbDataSchema } from '~/features/agicash-db/json-models';
-import { DestinationDetailsSchema } from '~/features/send/cashu-send-quote';
+import { DestinationDetailsSchema } from '~/features/shared/send-destination';
 import { TransactionStateSchema } from '../transaction-enums';
 import type { TransactionDetailsParserShape } from './transaction-details-types';
